fix datapath/control issue:
when b counts down to zero, there is 1 cycle delay in propogation of
next_state to DONE, which causes the adder to overflow by 1 cycle.
This was fixed by muxing the add operation with the cntz signal which
asserts when b counts down to zero. I don't find this elegant because it
moves some of the control to local singals instead of it all coming from
the controller.